requirements.txt
Django>=4.2.11,<5.0
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
Pillow>=10.4.0,<11.0
uwsgi>=2.0.24,<2.1
